refactor(step-ops): extract parsing utilities to lib/step-output.ts
Extracted pure utility functions from step-ops.ts (1202 -> 1058 lines):
- parseOutputKeyValues: parse KEY: value output format
- resolveTemplate: resolve {{placeholder}} templates
- findMissingTemplateKeys: find missing template variables

These functions have no DB or external dependencies, making them
suitable for extraction to the shared lib/ directory.

Backwards compatibility maintained via re-export in step-ops.ts.

Auto-generated by Openclaw AutoDev
